Factor out reducer for lesson plan data
