+++ client/app/bundles/course/lesson-plan/containers/LessonPlanFilter/index.jsx
@@ -88,5 +88,5 @@ class LessonPlanFilter extends React.Component {
 }
 
 export default connect(state => ({
-  visibility: state.visibilityByType,
+  visibility: state.lessonPlan.visibilityByType,
 }))(LessonPlanFilter);
