+++ client/app/bundles/course/lesson-plan/containers/LessonPlanNav/index.jsx
@@ -134,5 +134,5 @@ class LessonPlanNav extends React.Component {
 }
 
 export default connect(state => ({
-  groups: state.groups,
+  groups: state.lessonPlan.groups,
 }))(LessonPlanNav);
