+++ client/app/bundles/course/lesson-plan/pages/LessonPlanShow/LessonPlanItem/index.jsx
@@ -72,5 +72,5 @@ LessonPlanItem.propTypes = {
 };
 
 export default connect(state => ({
-  visibility: state.visibilityByType,
+  visibility: state.lessonPlan.visibilityByType,
 }))(LessonPlanItem);
