+++ client/app/bundles/course/lesson-plan/pages/LessonPlanShow/index.jsx
@@ -49,5 +49,5 @@ class LessonPlanShow extends React.Component {
 }
 
 export default connect(state => ({
-  groups: state.groups,
+  groups: state.lessonPlan.groups,
 }))(LessonPlanShow);
